+++ CHANGELOG.md
@@ -9,6 +9,9 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 
 ### Added
 
+- New link resolution style `:at_file_ref` which replaces links with the file path of the linked 
+  document in the form `@"path/to/file.md"`, that is used by many coding agents (Claude Code, 
+  Cursor etc.) to reference files  
 - Support for Elixir v1.16
 
 ### Changed